feed/google_news_rss/requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.4
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
feedgen==0.9.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gql==3.0.0a5
graphql-core==3.1.5
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
promise==2.3
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
six==1.16.0
toml==0.10.1
typing-extensions==3.10.0.0
urllib3==1.26.4
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.2.0
